+++ webapp/javascript/ui/LoadingSpinner.tsx
@@ -1,6 +1,4 @@
 import React from 'react';
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 import Spinner from 'react-svg-spinner';
 
 interface LoadingSpinnerProps {